Replace Bottle with Falcon, get a sane server setup (#8)
* Dirty push

* Switch to Falcon away from Bottle so that SparkBot has a real WSGI

* Use hmac.compare_digest() to protect against timing attacks

* Move some more code into the receiver module

* Update run.py to make things make a bit more sense

* Update inline documentation, remove useless code (removes help-all)

* Don't duplicate work.

The bot will always have an instance of spark_api. Because of this,
there is no need to make sure it has one.
Similarly, the bot will hold a webhook secret that the receiver can use.
No need to pull it in to a local variable.

* Don't require root_url for test reasons

* Update tests for switch to Falcon

* Warn about use of HTTP webhooks

* Documentation update (dirty push)

* Update documentation for Falcon receiver

* Dirty push

* Switch to Falcon away from Bottle so that SparkBot has a real WSGI

* Use hmac.compare_digest() to protect against timing attacks

* Move some more code into the receiver module

* Update run.py to make things make a bit more sense

* Update inline documentation, remove useless code (removes help-all)

* Don't duplicate work.

The bot will always have an instance of spark_api. Because of this,
there is no need to make sure it has one.
Similarly, the bot will hold a webhook secret that the receiver can use.
No need to pull it in to a local variable.

* Don't require root_url for test reasons

* Update tests for switch to Falcon

* Warn about use of HTTP webhooks

* Documentation update (dirty push)

* Update documentation for Falcon receiver

* Specify where run.py should be